solved mitm against tc02

+++ HW1/cipherImplementation.h
@@ -1,12 +1,12 @@
 #ifndef _CIPHERIMPLEMENTATION_H_
 #define _CIPHERIMPLEMENTATION_H_
-
+#include <stdint.h>
 // decleration for all functions and data types for cipherImplementation.c
 
 #define WORDSIZE 64  // size of one word in bits
 
 
-typedef unsigned long long int word;  // word is a block of TC01 encryption(and the key)
+typedef uint64_t word;  // word is a block of TC01 encryption(and the key)
 
 unsigned int extractNibble(word w, int i);
